Bug Fixes
Seed and crop price adjustments + additions

Milk buckets returned

+++ src/main/java/curtis1509/farmerslife/DepositCrop.java
@@ -5,14 +5,14 @@ import org.bukkit.Material;
 public class DepositCrop {
 
     Material material;
-    int reward;
+    double reward;
 
-    public DepositCrop(Material material, int reward){
+    public DepositCrop(Material material, double reward){
         this.material = material;
         this.reward = reward;
     }
 
-    public int getReward(){
+    public double getReward(){
         return reward;
     }
